try with stdio
