+++ test/PowerShellEditorServices.Test.E2E/TestsFixture.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Reflection;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using OmniSharp.Extensions.LanguageServer.Client;
+using OmniSharp.Extensions.LanguageServer.Client.Processes;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 using Xunit;
 
@@ -15,7 +17,28 @@ namespace PowerShellEditorServices.Test.E2E
         private readonly static string s_binDir =
             Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
 
-        public PowerShellEditorServicesProcess _psesProcess;
+        private readonly static string s_bundledModulePath = new FileInfo(Path.Combine(
+            s_binDir,
+            "..", "..", "..", "..", "..",
+            "module")).FullName;
+
+        private readonly static string s_sessionDetailsPath = Path.Combine(
+            s_binDir,
+            $"pses_test_sessiondetails_{Path.GetRandomFileName()}");
+
+
+        private readonly static string s_logPath = Path.Combine(
+            s_binDir,
+            $"pses_test_logs_{Path.GetRandomFileName()}");
+
+        const string s_logLevel = "Diagnostic";
+        readonly static string[] s_featureFlags = { "PSReadLine" };
+        const string s_hostName = "TestHost";
+        const string s_hostProfileId = "TestHost";
+        const string s_hostVersion = "1.0.0";
+        readonly static string[] s_additionalModules = { "PowerShellEditorServices.VSCode" };
+
+        private StdioServerProcess _psesProcess;
         public LanguageClient LanguageClient { get; private set; }
         public List<Diagnostic> Diagnostics { get; set; }
 
@@ -23,7 +46,39 @@ namespace PowerShellEditorServices.Test.E2E
         {
             var factory = new LoggerFactory();
 
-            _psesProcess = new PowerShellEditorServicesProcess(factory);
+            //_psesProcess = new PowerShellEditorServicesProcess(factory);
+
+            ProcessStartInfo processStartInfo = new ProcessStartInfo();
+            processStartInfo.FileName = Environment.GetEnvironmentVariable("PWSH_EXE_NAME") ?? "pwsh";
+            //_psesProcess.StartInfo.CreateNoWindow = true;
+            //_psesProcess.StartInfo.UseShellExecute = false;
+            //_psesProcess.StartInfo.RedirectStandardInput = true;
+            //_psesProcess.StartInfo.RedirectStandardOutput = true;
+            //_psesProcess.StartInfo.RedirectStandardError = true;
+            processStartInfo.ArgumentList.Add("-NoLogo");
+            processStartInfo.ArgumentList.Add("-NoProfile");
+            processStartInfo.ArgumentList.Add("-EncodedCommand");
+
+            string[] args = {
+                Path.Combine(s_bundledModulePath, "PowerShellEditorServices", "Start-EditorServices.ps1"),
+                "-LogPath", s_logPath,
+                "-LogLevel", s_logLevel,
+                "-SessionDetailsPath", s_sessionDetailsPath,
+                "-FeatureFlags", string.Join(',', s_featureFlags),
+                "-HostName", s_hostName,
+                "-HostProfileId", s_hostProfileId,
+                "-HostVersion", s_hostVersion,
+                "-AdditionalModules", string.Join(',', s_additionalModules),
+                "-BundledModulesPath", s_bundledModulePath,
+                "-Stdio"
+            };
+
+            var base64Str = System.Convert.ToBase64String(
+                System.Text.Encoding.Unicode.GetBytes(string.Join(' ', args)));
+
+            processStartInfo.ArgumentList.Add(base64Str);
+
+            _psesProcess = new StdioServerProcess(factory, processStartInfo);
             await _psesProcess.Start();
 
             LanguageClient = new LanguageClient(factory, _psesProcess);
